Refactoring DoubleBufferedAnimatedWindowHandler::render_frame().
Using a state machine to handle rendering the first two frames (and set
up the iterators) differently from the rest.

A lot of code could not be deduped due to the rust static analyzer, and
doing mutable borrows. Looked at using a RefCell to work around this,
but this added even more complicated/confusing/nasty type signatures
since the RefCell itself would have to be borrowed.