Heavily rework the subcommands.
Using traits and generic implementations of structs to standardize the
various creation of the clap subcommand descriptions, the parsing of
command line arguments, and running of each command.

Still want to make the subcommands return errors instead of
aporting/panicking, and possibly provide a list/map of the subcommands
for main to process them instead of having 2 module-level functions for
each subcommand.